refactor(scan): split ignore pattern owner
